% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{accelimp_list}
\alias{accelimp_list}
\title{Accelerometer Data Example with Imputations}
\format{
A list with 1 component:
\describe{
  \item{imp1}{A numeric matrix with 1,288 rows (daily profiles) and 1,440 columns
  (minute-by-minute accelerometer counts), containing the imputed accelerometer data.}
}
}
\source{
Data taken from the accelmissing package version 2.2.
}
\usage{
data(accelimp_list)
}
\description{
This dataset, accelimp_list, is a list containing imputed accelerometer data from the
2003-2004 National Health and Nutrition Examination Survey (NHANES). It includes
184 individuals, resulting in 1,288 daily profiles obtained after applying
\code{accel.impute()} to the raw accelerometer data.
}
\details{
The dataset name has been kept as 'accelimp_list' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
healthmotionR package and assists users in identifying its specific characteristics.
The suffix 'list' indicates that the dataset is stored as a list. The original content has not been modified
in any way.
}
