% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{angle_walk_array}
\alias{angle_walk_array}
\title{Hip and Knee Angle while Walking}
\format{
A 3-dimensional numeric array with 1,560 values and dimensions:
\describe{
  \item{[1:20]}{Movement cycle points}
  \item{[1:39]}{Individual subjects (boys)}
  \item{[1:2]}{Joint angle type: \code{"Hip Angle"}, \code{"Knee Angle"}}
}
}
\source{
Data taken from the \pkg{fda} package version 6.3.0.
}
\usage{
data(angle_walk_array)
}
\description{
This dataset, angle_walk_array, is a 3-dimensional array containing hip
and knee angle data (in degrees) for 39 boys measured during walking. Each observation
records the hip and knee joint angles across 20 equally spaced points of a movement cycle.
}
\details{
The dataset name has been kept as angle_walk_array to avoid confusion with other
datasets in the R ecosystem. This naming convention helps distinguish this dataset as
part of the healthmotionR package and assists users in identifying its specific
characteristics. The suffix _array indicates that the dataset is an array.
The original content has not been modified in any way.
}
