% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{AtrialFibrillation_list}
\alias{AtrialFibrillation_list}
\title{Atrial Fibrillation Multivariate Time Series}
\format{
A list with 2 components:
\describe{
  \item{data}{A list of 30 numeric matrices, each of dimension 640 × 2,
  representing two-channel ECG recordings.}
  \item{classes}{A numeric vector of length 30, indicating the class labels
  associated with each multivariate time series.}
}
}
\source{
Data taken from the \pkg{mlmts} package, version 1.1.2.
}
\usage{
data(AtrialFibrillation_list)
}
\description{
This dataset, AtrialFibrillation_list, is a multivariate time series (MTS)
consisting of two-channel ECG recordings of atrial fibrillation. The database
was created from data used in the Computers in Cardiology Challenge 2004.
}
\details{
The dataset name has been kept as 'AtrialFibrillation_list' to avoid confusion
with other datasets in the R ecosystem. This naming convention helps distinguish
this dataset as part of the healthmotionR package and assists users in identifying
its specific characteristics. The suffix 'list' indicates that the dataset is stored
as a list object. The original content has not been modified in any way.
}
