% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{dataset_info}
\alias{dataset_info}
\title{Get metadata for one dataset}
\usage{
dataset_info(name)
}
\arguments{
\item{name}{Dataset name or its language-neutral Python alias.}
}
\value{
A one-row data frame.
}
\description{
Get metadata for one dataset.
}
\examples{
dataset_info("body_metrics_df")
dataset_info("body_metrics")
}
