% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthmotionR-packages.R
\docType{package}
\name{healthmotionR}
\alias{healthmotionR}
\alias{healthmotionR-package}
\title{healthmotionR: A Comprehensive Collection of Health and Human Motion Datasets}
\description{
{
This package provides a broad collection of datasets focused on health, biomechanics, and human motion. It includes clinical, physiological, and kinematic information from diverse sources, covering aspects such as surgery outcomes, vital signs, rheumatoid arthritis, osteoarthritis, accelerometry, gait analysis, motion sensing, and biomechanics experiments.
}
}
\details{
healthmotionR: A Comprehensive Collection of Health and Human Motion Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Comprehensive Collection of Health and Human Motion Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/alejandrosialer/healthmotionR}
}
}
}
\author{
{
\strong{Maintainer}: Oscar Alejandro Sialer Gallo \email{alejandro.sialer.gallo@gmail.com}
}
}
