% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{infant_walking_df}
\alias{infant_walking_df}
\title{Infant Walking}
\format{
A data frame with 6 observations and 2 variables:
\describe{
  \item{exercise}{Numeric vector indicating the age at which infants in the exercise group began walking (months)}
  \item{no_exercise}{Numeric vector indicating the age at which infants in the no-exercise group began walking (months)}
}
}
\source{
Data taken from the \pkg{IIS} package version 1.1.
}
\usage{
data(infant_walking_df)
}
\description{
This dataset, infant_walking_df, is a data frame containing the ages (in months)
at which 12 infants were reported by their mothers to have started walking.
The infants were randomly assigned to either an "exercise" or "no-exercise" group
as part of the study conducted by Zelazo et al. (1972). The data are also presented
in Table 9.8 of Wolfe and Schneider, *Intuitive Introductory Statistics*.
}
\details{
The dataset name has been kept as 'infant_walking_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
healthmotionR package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
