% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{z_labels_monitoring_df}
\alias{z_labels_monitoring_df}
\title{Activity Labels for Human Activity Monitoring}
\format{
A data frame with 6 observations and 3 variables:
\describe{
  \item{activity}{Character string indicating the recorded activity}
  \item{start}{Integer value indicating the start time of the activity}
  \item{end}{Integer value indicating the end time of the activity}
}
}
\source{
Data taken from the TrendLSW package version 1.0.2.
}
\usage{
data(z_labels_monitoring_df)
}
\description{
This dataset, z_labels_monitoring_df, is a data frame containing the labelled activities
recorded during the observation period corresponding to the data object z.acc.
}
\details{
The dataset name has been kept as 'z_labels_monitoring_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
healthmotionR package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
